pyproject.toml
[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[project]
name = "labelguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.5.1",
    "hydra-core>=1.3.2",
    "mlflow>=3.10.1",
    "openai>=2.30.0",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "qdrant-client==1.14.1",
    "s3fs>=2026.2.0",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "xgboost>=3.2.0",
]
